Remove circular dependencies between actions/models for QuerySelect
